Keep IME regression coverage reviewable
The Zhuyin marked-text regression tests and helper logic were valid, but they pushed already-large files over the review budget. Move the marked-text range helper into a focused GhosttyNSView extension and keep the new IME coverage in a dedicated test file so the behavior stays covered without expanding the overloaded CJK test file.

Constraint: CodeRabbit file budget review requires both GhosttyTerminalView.swift and CJKIMEInputTests.swift to stay under their configured limits

Rejected: Leave the helpers inline | it keeps the behavioral fix correct but fails the repository review budget

Confidence: high

Scope-risk: narrow

Directive: Keep active marked-text range helpers near IME composition code instead of re-expanding GhosttyTerminalView.swift

Tested: ./scripts/reload.sh --tag issue-3571-zhuyin-ime-candidate

Not-tested: Local XCTest per project testing policy
